services/recording/requirements.txt
# Recording Service Dependencies

# Web framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0

# WebSocket client
websockets>=12.0,<13.0

# Data storage
pyarrow>=15.0.0,<16.0.0

# Configuration
pyyaml>=6.0.1,<7.0.0

# Type hints and validation
pydantic>=2.5.0,<3.0.0

# Testing (dev only)
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.27.0,<1.0.0  # For FastAPI TestClient
